fix: move users to accounts
- move users to accounts
- use django user model as base model to use later in authentication
